Fix broken Fetch Data Svelte 5 component
Also simplify the code by using state fields.
